requirements.txt
aiogram==3.4.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
aiosqlite==0.19.0
httpx==0.26.0

